Localization files, and goto command
Also adjusted player/find.lua to have more error resiliency. Of course, with the localization files being added, I also localized some internals.
Developing in singleplayer can be good for a more powerful auto-reload, but unfortunately the console is the server's console. This means I am running the server sided master command instead of the networked client command. When in singleplayer, there is now the "sv_pyrition" command as a server-sided variant to the "pyrition" command. I may have to make one of these for listen servers too.